requirements.txt
tensorflow-cpu
numpy
librosa
scikit-learn
matplotlib
seaborn
pandas
soundfile
streamlit
flask
werkzeug